backend/requirements.txt
fastapi
uvicorn
python-multipart
sentence-transformers
faiss-cpu
requests
transformers
pydub
whisper
